frontend/requirements.txt
streamlit
streamlit-option-menu
requests
pandas
markdown
python-dotenv
PyYAML